+++ lib/PoweroidSdk10/pin_io.h
@@ -1,10 +1,12 @@
 #ifndef PIN_IO_H
 #define PIN_IO_H
 
-bool readPinLow(int pin);
-bool readPinHigh(int pin);
-int readPinVal(int pin);
-void led(int no, bool on);
-void pin_inv(int no, bool on);
+#include <Arduino.h>
+
+bool readPinLow(uint8_t pin);
+bool readPinHigh(uint8_t pin);
+int readPinVal(uint8_t pin);
+void led(uint8_t no, bool on);
+void pin_inv(uint8_t no, bool on);
 
 #endif
